requirements_all_branches.txt
# 合并所有分支的依赖
# 自动生成于: 2025-12-11
# 支持所有6个分支运行

# ============================================================================
# LangChain 核心
# ============================================================================
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# ============================================================================
# Google Gemini
# ============================================================================
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
google-genai>=0.2.0  # Official SDK for grounding with Google Search

# ============================================================================
# 基础依赖
# ============================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0

# ============================================================================
# 搜索工具（可选）
# ============================================================================
google-search-results>=2.4.2  # SerpAPI
google-api-python-client>=2.0.0  # Google Custom Search

# ============================================================================
# RAG 相关（仅 baseline-rag 分支需要）
# ============================================================================
faiss-cpu>=1.7.4  # 向量检索
numpy>=1.24.0  # 数值计算

# ============================================================================
# 数据处理
# ============================================================================
pandas>=2.0.0
matplotlib>=3.7.0

# ============================================================================
# 开发工具
# ============================================================================
ipython>=8.0.0

# ============================================================================
# 总计：17个依赖包
# 说明：
# - 所有分支共享大部分依赖（LangChain, Gemini, 基础工具）
# - faiss-cpu 和 numpy 仅 RAG 分支需要，但安装了也不影响其他分支
# - 使用 >= 版本约束，确保兼容性
# ============================================================================
